[WIP] An example of case, when multithreading causes trouble
It's too bad, but there is no guarantee of data uniqueness - two threads can process copies of the same pictures in the same folder and because they check DB concurrently, both results (identical) will be stored. This bug appears only when there are copies of the same picture processed by two or more threads simultaneously.